+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 * add optional `screenCalibration` in `LocalContext`
 
+### Bug Fixes
+
+* remove leftover `Etherpad` type export from SDK entrypoint
+
 ## [5.5.0](https://github.com/graasp/graasp-sdk/compare/v5.4.0...v5.5.0) (2024-12-12)